Change volume filesystem total and free to GB
Previously filesystsem total and free attributes were returned as
block sizes (ints).  Now they will be returned as GB size as floats.

+++ src/nova/guest/diagnostics.h
@@ -44,8 +44,8 @@ namespace nova { namespace guest { namespace diagnostics {
         unsigned long block_size;
         unsigned long total_blocks;
         unsigned long free_blocks;
-        unsigned long total;
-        unsigned long free;
+        float total;
+        float free;
         float used;
     };
